lifting state up to add book page
